rename to BrowserAndInputScreenTransitionProviderImpl and use DI
