+++ duckchat/duckchat-impl/src/main/java/com/duckduckgo/duckchat/impl/inputscreen/ui/InputScreenActivity.kt
@@ -21,9 +21,8 @@ import android.os.Bundle
 import com.duckduckgo.anvil.annotations.ContributeToActivityStarter
 import com.duckduckgo.anvil.annotations.InjectWith
 import com.duckduckgo.common.ui.DuckDuckGoActivity
-import com.duckduckgo.common.ui.store.BrowserAppTheme
 import com.duckduckgo.di.scopes.ActivityScope
-import com.duckduckgo.duckchat.api.InputScreenAnimationResourceProvider
+import com.duckduckgo.duckchat.api.inputscreen.BrowserAndInputScreenTransitionProvider
 import com.duckduckgo.duckchat.impl.R
 import com.duckduckgo.navigation.api.GlobalActivityStarter
 import javax.inject.Inject
@@ -37,7 +36,7 @@ data class InputScreenActivityParams(
 class InputScreenActivity : DuckDuckGoActivity() {
 
     @Inject
-    lateinit var appTheme: BrowserAppTheme
+    lateinit var browserAndInputScreenTransitionProvider: BrowserAndInputScreenTransitionProvider
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -50,8 +49,8 @@ class InputScreenActivity : DuckDuckGoActivity() {
     }
 
     private fun applyExitTransition() {
-        val enterTransition = InputScreenAnimationResourceProvider.getBrowserEnterAnimation(appTheme.isLightModeEnabled())
-        val exitTransition = InputScreenAnimationResourceProvider.getInputScreenExitAnimation()
+        val enterTransition = browserAndInputScreenTransitionProvider.getBrowserEnterAnimation()
+        val exitTransition = browserAndInputScreenTransitionProvider.getInputScreenExitAnimation()
 
         if (VERSION.SDK_INT >= 34) {
             overrideActivityTransition(
